+++ src/napari/settings/_base.py
@@ -4,7 +4,6 @@ import contextlib
 import json
 import logging
 import os
-import re
 from collections.abc import Mapping
 from enum import StrEnum
 from functools import partial
@@ -19,7 +18,6 @@ from pydantic import (
     PrivateAttr,
     TypeAdapter,
     ValidationError,
-    create_model,
 )
 from pydantic_settings import (
     BaseSettings,
@@ -29,7 +27,6 @@ from pydantic_settings import (
 )
 
 from napari._pydantic_util import get_inner_type, get_origin
-from napari.settings._plugin_preferences import PluginPreferences
 from napari.settings._yaml import PydanticYamlMixin
 from napari.utils.events import EmitterGroup, EventedModel
 from napari.utils.misc import StringEnum, deep_update
@@ -507,119 +504,3 @@ def _remove_empty_dicts(dct: dict, recurse=True) -> dict:
         if v == {}:
             del dct[k]
     return dct
-
-
-def _field_name(key: str, plugin_name: str) -> str:
-    """
-    Convert:
-        my_plugin.someSetting -> some_setting
-    """
-    key = key.removeprefix(f'{plugin_name}.')
-
-    key = re.sub(
-        r'(?<!^)(?=[A-Z])',
-        '_',
-        key,
-    )
-
-    key = re.sub(
-        r'[.\-\s]+',
-        '_',
-        key,
-    )
-
-    return key.lower()
-
-
-VALUE_TRANSLATOR = {
-    'maximum': 'le',
-    'minimum': 'ge',
-    'exclusive_maximum': 'lt',
-    'exclusive_minimum': 'gt',
-}
-_TYPE_MAP: dict[str, type] = {
-    'boolean': bool,
-    'string': str,
-    'integer': int,
-    'number': float,
-    'array': list,
-    'int': int,
-    'float': float,
-    'str': str,
-    'bool': bool,
-    'list': list,
-}
-
-
-def _build_single_config_model(
-    configuration,
-    plugin_name: str,
-) -> type[EventedModel]:
-
-    fields = {}
-
-    for key, props in configuration.properties.items():
-        if props.type is None:
-            continue
-
-        data = {k: getattr(props, k) for k in props.model_fields_set}
-
-        type_name = data.pop('type')
-        field_type = _TYPE_MAP.get(type_name)
-
-        field_kwargs = {VALUE_TRANSLATOR.get(k, k): v for k, v in data.items()}
-
-        field_name = _field_name(key, plugin_name)
-
-        fields[field_name] = (
-            field_type,
-            Field(**field_kwargs),
-        )
-    model_name = configuration.title.lower()
-    model = create_model(
-        model_name,
-        __base__=EventedModel,
-        **fields,
-    )
-    model.display = configuration.title
-    return model
-
-
-def plugin_configuration_generator() -> dict[str, PluginPreferences]:
-    from npe2 import PluginManager
-
-    pm = PluginManager.instance()
-    pm.discover()
-    plugins = sorted(
-        pm.iter_manifests(),
-        key=lambda x: x.name,
-    )
-    display_names = {plugin.name: plugin.display_name for plugin in plugins}
-    plugin_contr = {
-        plug.name: plug.contributions for plug in plugins if plug.contributions
-    }
-    configurations = {
-        plug: conf.configuration
-        for plug, conf in plugin_contr.items()
-        if conf.configuration
-    }
-    plugin_settings = {}
-    for plugin_name, configuration in configurations.items():
-        models = [
-            _build_single_config_model(conf, plugin_name)
-            for conf in configuration
-        ]
-        fields = {}
-
-        for model in models:
-            fields[model.__name__] = (
-                model,
-                Field(default_factory=model, title=model.display),
-            )
-        plugin_settings[plugin_name] = create_model(
-            f'{plugin_name} Preferences',
-            __base__=PluginPreferences,
-            **fields,
-        )
-        plugin_settings[plugin_name].display_name = display_names[plugin_name]
-    return plugin_settings
